requirements.txt
Flask>=3.0.0
transformers>=4.36.0
torch>=2.1.0
Pillow>=10.0.0
accelerate>=0.24.0
